Add Users page and related functionalities
Added a new Users page in the Pages directory. This includes the related components and functionalities like user options and profile pictures. Also, included the option to navigate to the Users page from the sidebar navigation. Made some changes to the login page visuals and wording. Also, added 'assets' folder in the 'LoginBlazor.csproj'.