Add bootstrap.css containing sizing and general styling
A new bootstrap.css file is added that contains general styling information, including the settings for the sizing of various elements. With this file, we can standardize the appearance of components across the application. Also, the bootstrap version provided enhances the consistency and scalability of the project by providing a uniform way to style the elements.