Update UI components and add delete user role function
This commit refactors multiple components in different pages. The main change introduced is the delete functionality in the "Rolesporusuario" page for user roles. Minor updates include changing the profile picture path in "MainMenu", updating the error handling in "Users", and adding new table representation in "SucursalesRoles". Additionally, the project dependencies are updated with the addition of "Microsoft.EntityFrameworkCore".